refactor: centralise css logic
